+++ BankManagerConsoleApp/BankManagerConsoleApp/PriorityQueue.swift
@@ -114,6 +114,7 @@ struct PriorityQueue<Element>: Queue {
     
     var peek: Element? { heap.peek() }
     
+    @discardableResult
     mutating func enqueue(_ element: Element) -> Bool {
         heap.insert(element)
         return true
